マルチユーザー対応できた。認証周りよくわからなくてあと5000回くらい書かないと理解できなさそう。

+++ backend/routes/todoRoutes.js
@@ -1,15 +1,18 @@
 const express = require('express');
 const Todo = require('../models/Todo');
-
+const authMiddleware = require('../middleware/authMiddleware');
+require('dotenv').config();
 const router = express.Router();
+
+router.use(authMiddleware);
 
 // todo取得
 router.get('/', async (req, res) => {
     try {
-        const todos = await Todo.find();
+        const todos = await Todo.find({ userId: req.user._id });
         res.json(todos);
     } catch (error) {
-        console.error("Error! Fetching todos: ", error);
+        console.error("Error! Fetching todos: ", error.message);
         res.status(500).json({message: "Server Error"});
     }
 });
@@ -20,6 +23,7 @@ router.post('/', async (req, res) => {
         const newTodo = new Todo({
             name: req.body.name,
             completed: false,
+            userId: req.user._id
         });
         await newTodo.save();
         res.json(newTodo);
@@ -33,7 +37,7 @@ router.post('/', async (req, res) => {
 router.put('/:id', async (req, res) => {
     try {
         const updatedTodo = await Todo.findByIdAndUpdate(
-            req.params.id,
+            { _id:req.params.id, userId: req.user._id},
             { completed: req.body.completed, name: req.body.name },
             { new: true } 
         );
@@ -47,7 +51,7 @@ router.put('/:id', async (req, res) => {
 // todo削除
 router.delete('/:id', async (req, res) => {
     try {
-        const deletedTodo = await Todo.findByIdAndDelete(req.params.id);
+        const deletedTodo = await Todo.findByIdAndDelete({ _id: req.params.id, userId: req.user._id });
         res.json(deletedTodo)
     } catch (error) {
         console.error("Error! Deleting todo: ", error);
